Extract the pending-song commit path into a shared edge-function module
auto-commit-song owned the only copy of "turn a pending_songs row into
files in works/". A retry pass needs the same code, and a second copy of
it would drift from this one the first time either changed.

_shared/commit-song.ts now holds the GitHub contents-API calls, the YAML
builders, and commitPendingSong()/commitAttachment(). auto-commit-song
keeps its auth, validation, submission_log, and response shape unchanged
and calls into the module.

Two behaviour notes:
- getFileContent() distinguishes a real 404 from a transient GitHub
  error. The old attachment path treated any failed read as "no
  work.yaml", which with create_placeholder set would overwrite a real
  work.yaml with a placeholder.
- unretryableReason() names the rows no retry can fix (no id, no title,
  no content) so the reconciler can report them instead of hammering
  GitHub with them every hour.

Supabase bundles relative imports, so both functions must be redeployed
together when this module changes.
